% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S5Processor.R
\docType{package}
\name{S5Processor}
\alias{S5Processor}
\alias{S5Processor-package}
\title{Sentinel-5P L2 Processing Tool}
\description{
This R package automatically converts the Sentinel-5P L2 NetCDF files into TIFF files.
The user can define the product which should be turned into a TIFF and also mosaic from
a list of file-paths.
}
\details{
Function included in this package:
\itemize{
  \item \code{\link{S5P_process}}
}
}
\references{
This program was written at Remote Sensing Solutions RSS GmbH,
            Dingolfinger Str. 9, 81673 Munich, Germany \url{https://rssgmbh.de/}.
}
\author{
Marius Philipp \email{marius.philipp@stud-mail.uni-wuerzburg.de}
}
